app/requirements.txt
absl-py==1.4.0
amqp==5.1.1
anyio==3.6.2
astunparse==1.6.3
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
celery-amqp-backend==1.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
contours==0.0.2
cycler==0.11.0
dicom==0.9.9.post1
fastapi==0.89.1
flatbuffers==23.1.21
flower==1.2.0
fonttools==4.38.0
future==0.18.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.6.0
idna==3.4
install==1.3.5
joblib==1.2.0
# keras==2.10.0
# Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.2.4
libclang==15.0.6.1
lightgbm==3.3.5
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.24.1
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
postgres==4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pydicom==2.3.1
PyJWT==2.6.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
python-version==0.0.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.1
scikit-optimize==0.8.1
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.22.0
# tensorboard==2.10.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
# tensorflow-estimator==2.10.0
# tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.7.3
supervisor==4.1.0